proper separation of whitebox/blackbox macros
