+++ chimney/src/main/scala/io/scalaland/chimney/Transformer.scala
@@ -1,6 +1,6 @@
 package io.scalaland.chimney
 
-import io.scalaland.chimney.internal.ChimneyMacros
+import io.scalaland.chimney.internal.ChimneyBlackboxMacros
 import scala.language.experimental.macros
 
 trait Transformer[From, To] {
@@ -11,5 +11,5 @@ trait Transformer[From, To] {
 object Transformer {
 
   implicit def derive[From, To]: Transformer[From, To] =
-    macro ChimneyMacros.deriveTransformerImpl[From, To]
+    macro ChimneyBlackboxMacros.deriveTransformerImpl[From, To]
 }
