+++ chimney/src/main/scala/io/scalaland/chimney/internal/DerivationGuards.scala
@@ -1,11 +1,11 @@
 package io.scalaland.chimney.internal
 
-import scala.reflect.macros.whitebox
+import scala.reflect.macros.blackbox
 
 trait DerivationGuards {
   this: MacroUtils =>
 
-  val c: whitebox.Context
+  val c: blackbox.Context
 
   import c.universe._
